POINT AND CLICK CONTROLS!!!!!!!
This took a lot of time to figure out but I finally got it working. Basically what its doing is shooting out a 2D cast, checking if theres a collision, adding that collision to the appropriate queue and/or moving towards there. Then interacting as appropriately.

I am really happy this finally works!

Additionally, I added the outline of objects for when you hover with the mouse as well, not just when you walk in range. This allows the player to know an object is interactable just by hovering it.